Add a monostate type
This is equivalent to std::monostate, but avoids including "variant"
just for that.

The patch maps void to monostate in a few places, which simplifies the
code since it is valid to have a reference to monostate, but not to
void.

This will be particularly useful in changing the loop future to
produce a value.

Signed-off-by: Rafael Ávila de Espíndola <rafael@espindo.la>