Remove posix_error::ok()
While it is common in c++ for error types to include a "no error"
value, it is a bit confusing when we also have a result type.

This moves us closer to Rust and uses a result for the return of
functions that may or may not fail, but don't produce any value.

Signed-off-by: Rafael Ávila de Espíndola <rafael@espindo.la>