Add a dedicated and_then_future
The advantage over implementing on top of then_future is that
then_future::poll is

  if (!call_done) {
    // call this->func
  }
  return after.poll(react);

In this case, func was effectively internal::and_then_impl, which
returns either<ready_future, X>.

We know that once a ready value is returned, there will be no followup
calls to poll. The compiler doesn't.

We could specialize then_future for when func returns a ready_future or
an either<ready_future, X>.

The ready_future specialization would be similar to the current
specialization for func returning a non-future value.

The either<ready_future, X> specialization could store just the future
X in 'after'.

For now it is simpler to just implement and_then directly. It also
makes fewer assumptions on how much the compiler can handle.

Signed-off-by: Rafael Ávila de Espíndola <rafael@espindo.la>