Optimize future::and_then
When the function object passed to future::and_then returns a value
instead of a future, it will now pass a lambda to future::then that
also returns a value.

Doing that means than now future::and_then also benefits from the
optimization done on the previous patch to future::then.

That also removes the last use of futurize::apply, so it is removed.

Signed-off-by: Rafael Ávila de Espíndola <rafael@espindo.la>